+++ wifi/pxt.json
@@ -8,6 +8,7 @@
     ],
     "dependencies": {
         "core": "*",
-        "jacdac": "github:microsoft/pxt-jacdac-services"
+        "jacdac": "github:microsoft/pxt-jacdac-services",
+        "jacdac-tcp": "github:microsoft/pxt-jacdac-services/tcp"
     }
 }
